+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [Unreleased]
+
+### Added
+
+### Changed
+
+### Removed
+
+### Fixed
+
+
+
+
+
+
 ## [v2.4.1] 2025-08-25
 
 ### Fixed